MinimalRNNCell now doesn't give NaNs
I added an tanh activation afterwards and changed the weight
initializations. These three defaults were copied from SimpleRNN, which
seemed to work. So, basically 'MinimalRNNCell' is no longer minimal and
performs just as well as SimpleRNN, but it's a test at implementing my
own RNN cell in a Python class.