Train domain additional time
This seems to have a great effect. Though, it aligns S0 with T1 and S1 with T0, but
I don't think with this algorithm as-is be able to properly align those
since we can't use the labels. (Maybe something like a semantic
consistency loss like in CyCADA though.) The t-SNE and PCA look much
better aligned by doing this.